Platinum_Tier/requirements.txt
# Platinum Tier Requirements
# Enterprise AI Employee System

# Core Dependencies
asyncio-mqtt>=0.16.1
aiohttp>=3.9.0
aiofiles>=23.2.1

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Database
redis>=5.0.1
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25

# Vector Database
# Choose one:
pinecone-client>=3.0.0  # Cloud option
chromadb>=0.4.22  # Self-hosted option

# Embeddings
sentence-transformers>=2.3.1
transformers>=4.37.0

# Voice Integration
# Vapi (via API calls - no package needed)
# Retell AI (via API calls - no package needed)
requests>=2.31.0

# Security
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Monitoring
prometheus-client>=0.19.0
python-json-logger>=2.0.7

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.7
rich>=13.7.0

# Email
google-api-python-client>=2.115.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Social Media
tweepy>=4.14.0
facebook-sdk>=3.1.0
instagrapi>=2.0.0

# Odoo
odoorpc>=0.10.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# Development
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0

# Watchdog for file monitoring
watchdog>=4.0.0

# Additional utilities
schedule>=1.2.0
pytz>=2024.1
